#79 factor out *ColumnConstraint message creation
