+++ src/main/scala/de/frosner/ddq/constraints/StringColumnConstraint.scala
@@ -23,20 +23,12 @@ case class StringColumnConstraintResult(constraint: StringColumnConstraint,
                                         data: Option[StringColumnConstraintResultData],
                                         status: ConstraintStatus) extends ConstraintResult[StringColumnConstraint] {
 
-  val message: String = {
-    val constraintString = constraint.constraintString
-    val maybeViolatingRows = data.map(_.failedRows)
-    val maybePluralS = maybeViolatingRows.map(violatingRows => if (violatingRows == 1) "" else "s")
-    (status, maybeViolatingRows, maybePluralS) match {
-      case (ConstraintSuccess, Some(0), _) =>
-        s"Constraint $constraintString is satisfied."
-      case (ConstraintFailure, Some(violatingRows), Some(pluralS)) =>
-        s"$violatingRows row$pluralS did not satisfy constraint $constraintString."
-      case (ConstraintError(throwable), None, None) =>
-        s"Checking constraint $constraintString failed: $throwable"
-      case default => throw IllegalConstraintResultException(this)
-    }
-  }
+  val message: String = ColumnConstraintUtil.createColumnConstraintMessage(
+    status = status,
+    constraintResult = this,
+    constraintString = constraint.constraintString,
+    maybeViolatingRows = data.map(_.failedRows)
+  )
 
 }
